Refactor: Move products to own directory
