+++ src/Catalog/Product/ProductFixturePool.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TddWizard\Fixtures\Catalog;
+namespace TddWizard\Fixtures\Catalog\Product;
 
 use Magento\Catalog\Api\Data\ProductInterface;
 use function array_values as values;
