+++ tests/Checkout/CartBuilderTest.php
@@ -6,8 +6,8 @@ namespace TddWizard\Fixtures\Checkout;
 
 use Magento\Quote\Model\Quote\Item;
 use PHPUnit\Framework\TestCase;
-use TddWizard\Fixtures\Catalog\ProductBuilder;
-use TddWizard\Fixtures\Catalog\ProductFixture;
+use TddWizard\Fixtures\Catalog\Product\ProductBuilder;
+use TddWizard\Fixtures\Catalog\Product\ProductFixture;
 
 class CartBuilderTest extends TestCase
 {
@@ -16,7 +16,7 @@ class CartBuilderTest extends TestCase
     protected function setUp(): void
     {
         $this->productFixture = new ProductFixture(
-            ProductBuilder::aSimpleProduct()->build()
+            ProductBuilder::aSimpleProduct()->build(),
         );
     }
 
@@ -32,7 +32,7 @@ class CartBuilderTest extends TestCase
         $cart = CartBuilder::forCurrentSession()->withProductRequest(
             $this->productFixture->getSku(),
             $qty,
-            ['options' => [$customOptionId => $customOptionValue]]
+            ['options' => [$customOptionId => $customOptionValue]],
         )->build();
         $quoteItems = $cart->getQuote()->getAllItems();
         $this->assertCount(1, $quoteItems, "1 quote item should be added");
@@ -46,7 +46,7 @@ class CartBuilderTest extends TestCase
         $this->assertJsonStringEqualsJsonString(
             json_encode(['qty' => $qty, 'options' => ['42' => 'foobar']]),
             $serializedBuyRequest,
-            "Value of info_buyRequest option should be as configured"
+            "Value of info_buyRequest option should be as configured",
         );
     }
 }
