+++ tests/Checkout/CustomerCheckoutTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace TddWizard\Fixtures\Checkout;
 
 use PHPUnit\Framework\TestCase;
-use TddWizard\Fixtures\Catalog\ProductBuilder;
-use TddWizard\Fixtures\Catalog\ProductFixturePool;
+use TddWizard\Fixtures\Catalog\Product\ProductBuilder;
+use TddWizard\Fixtures\Catalog\Product\ProductFixturePool;
 use TddWizard\Fixtures\Customer\AddressBuilder;
 use TddWizard\Fixtures\Customer\CustomerBuilder;
 use TddWizard\Fixtures\Customer\CustomerFixturePool;
@@ -25,16 +25,16 @@ class CustomerCheckoutTest extends TestCase
         $this->customerFixtures = new CustomerFixturePool();
         $this->customerFixtures->add(
             CustomerBuilder::aCustomer()->withAddresses(
-                AddressBuilder::anAddress()->asDefaultBilling()->asDefaultShipping()
-            )->build()
+                AddressBuilder::anAddress()->asDefaultBilling()->asDefaultShipping(),
+            )->build(),
         );
         $this->productFixtures->add(
             ProductBuilder::aSimpleProduct()->withPrice(10)->build(),
-            'simple'
+            'simple',
         );
         $this->productFixtures->add(
             ProductBuilder::aVirtualProduct()->withPrice(10)->build(),
-            'virtual'
+            'virtual',
         );
     }
 
@@ -53,8 +53,8 @@ class CustomerCheckoutTest extends TestCase
         $this->customerFixtures->get()->login();
         $checkout = CustomerCheckout::fromCart(
             CartBuilder::forCurrentSession()->withSimpleProduct(
-                $this->productFixtures->get('simple')->getSku()
-            )->build()
+                $this->productFixtures->get('simple')->getSku(),
+            )->build(),
         );
         $order = $checkout->placeOrder();
         $this->assertNotEmpty($order->getEntityId(), 'Order should be saved successfully');
@@ -70,14 +70,14 @@ class CustomerCheckoutTest extends TestCase
         $this->customerFixtures->get()->login();
         $checkout = CustomerCheckout::fromCart(
             CartBuilder::forCurrentSession()->withSimpleProduct(
-                $this->productFixtures->get('virtual')->getSku()
-            )->build()
+                $this->productFixtures->get('virtual')->getSku(),
+            )->build(),
         );
         $order = $checkout->placeOrder();
         $this->assertNotEmpty($order->getEntityId(), 'Order should be saved successfully');
         $this->assertEmpty(
             $order->getExtensionAttributes()->getShippingAssignments(),
-            'Order with virtual product should not have any shipping assignments'
+            'Order with virtual product should not have any shipping assignments',
         );
         $this->assertEmpty($order->getShippingDescription(), 'Order should not have a shipping description');
     }
